+++ Helpers/P5/Models/Line.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 namespace ExoKomodo.Helpers.P5.Models
 {
     public struct Line
@@ -6,12 +8,12 @@ namespace ExoKomodo.Helpers.P5.Models
 
         #region Constructors
         public Line(
-            double x1 = 0,
-            double y1 = 0,
-            double z1 = 0,
-            double x2 = 0,
-            double y2 = 0,
-            double z2 = 0
+            float x1 = 0,
+            float y1 = 0,
+            float z1 = 0,
+            float x2 = 0,
+            float y2 = 0,
+            float z2 = 0
         ) : this(new Vector3(x1, y1, z1), new Vector3(x2, y2, z2)) {}
 
         public Line(Vector3 start, Vector3 end)
